requirements.txt
PyYAML==3.10
pytest==2.3.5
